@@ -4,25 +4,30 @@
   "private": true,
   "license": "MIT",
   "dependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "change-case": "4.1.2",
-    "firebase": "9.1.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-router": "5.2.1",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.3",
-    "semantic-ui-react": "2.0.4",
-    "styled-components": "5.3.8"
+    "firebase": "10.2.0",
+    "postcss": "^8.4.28",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router": "6.15.0",
+    "react-router-dom": "6.15.0",
+    "react-scripts": "5.0.1",
+    "semantic-ui-react": "2.1.4",
+    "styled-components": "6.0.7"
   },
   "devDependencies": {
-    "firebase-tools": "^9.20.0",
-    "husky": "7.0.2",
+    "firebase-tools": "12.4.8",
+    "husky": "8.0.3",
     "jsonlint": "1.6.3",
-    "lint-staged": "11.2.3",
-    "prettier": "2.8.4"
+    "lint-staged": "14.0.0",
+    "prettier": "3.0.2"
+  },
+  "peerDependencies": {
+    "react-is": "16.8.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "deploy": "firebase deploy",
