@@ -1,9 +1,23 @@
-import React, { Component } from 'react';
-import { BrowserRouter as Router, Route } from "react-router-dom";
+import React, { Component } from "react";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Navigate,
+  Outlet,
+} from "react-router-dom";
 import styled from "styled-components";
-import firebase from "firebase/app";
 
-import { Navbar } from './Navbar';
+import { initializeApp } from "firebase/app";
+import {
+  getFirestore,
+  collection,
+  getDocs,
+  doc,
+  setDoc,
+} from "firebase/firestore/lite";
+
+import { Navbar } from "./Navbar";
 import { Parks } from "./Parks";
 import { Home } from "./Home";
 import { Park } from "./Park";
@@ -11,7 +25,6 @@ import { Profile } from "./Profile";
 import { Signin } from "./Signin";
 import { Signup } from "./Signup";
 import { Login } from "./Login";
-import { PrivateRoute } from "./PrivateRoute";
 // import { AuthButton } from "./AuthButton";
 
 // Initialize Firebase
@@ -21,30 +34,41 @@ var config = {
   databaseURL: "https://footprints-442fb.firebaseio.com",
   projectId: "footprints-442fb",
   storageBucket: "",
-  messagingSenderId: "576698370523"
+  messagingSenderId: "576698370523",
 };
-firebase.initializeApp(config);
+const app = initializeApp(config);
 
 const StyledApp = styled.div`
   text-align: center;
 `;
+
+const PrivateRoute = () => {
+  const fakeAuth = {
+    isAuthenticated: false,
+    authenticate(cb) {
+      this.isAuthenticated = true;
+    },
+    signout(cb) {
+      this.isAuthenticated = false;
+    },
+  };
+  // If authorized, return an outlet that will render child elements
+  // If not, return element that will navigate to login page
+  return fakeAuth ? <Outlet /> : <Navigate to="/login" />;
+};
 export class App extends Component {
   state = {};
 
-  _fetchParks() {
-    const parks = firebase.database().ref("parks");
-    return parks.on("value", function(snapshot) {
-      console.log(snapshot.val());
-    });
+  async _fetchParks() {
+    let parks = await getDocs(collection(getFirestore(app), "parks"));
+    return parks;
   }
 
-  _storeHighScore(userId, score) {
-    firebase
-      .database()
-      .ref("users/" + userId)
-      .set({
-        highscore: score
-      });
+  async _storeHighScore(userId, score) {
+    const ref = doc(getFirestore(app), "users/" + userId);
+    await setDoc(ref, {
+      highscore: score,
+    });
   }
 
   render() {
@@ -57,14 +81,16 @@ export class App extends Component {
       <Router>
         <Navbar />
         <StyledApp>
-          <Route path="parks/:name" component={Park} />
-          {/* <AuthButton /> */}
-          <Route path="/login" component={Login} />
-          <Route path="/signin" component={Signin} />
-          <Route path="/signup" component={Signup} />
-          <Route path="/parks" component={Parks} />
-          <PrivateRoute path="/me" component={Profile} />
-          <Route path="/" component={Home} activeItem={activeItem}/>
+          <Routes>
+            <Route path="parks/:name" element={<Park/>} />
+            {/* <AuthButton /> */}
+            <Route path="/login" element={<Login/>} />
+            <Route path="/signin" element={<Signin/>} />
+            <Route path="/signup" element={<Signup/>} />
+            <Route path="/parks" element={<Parks/>} />
+            <Route path="/me" element={<PrivateRoute><Profile/></PrivateRoute>}/>
+            <Route path="/" element={<Home/>} activeItem={activeItem} />
+          </Routes>
         </StyledApp>
       </Router>
     );
